% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{add_grid_to_map}
\alias{add_grid_to_map}
\title{Add Grid to Map}
\usage{
add_grid_to_map(my_map)
}
\arguments{
\item{my_map}{An existing leaflet map.}
}
\value{
The updated leaflet map with added grid.
}
\description{
This function adds a grid (graticule) to an existing leaflet map.
}
\details{
This function uses the `clearGroup` and `addSimpleGraticule` functions 
from the `leaflet` package to add a grid to the map. The grid is defined by the 
specified interval, and it includes latitude and longitude lines. The grid is 
added to the "Grid" group on the map.
}
\examples{
\dontrun{
# Example usage
my_map <- init_map_of_sensors(sensors_data)
my_map <- add_grid_to_map(my_map)
}
}
