% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{add_grid_to_map_without_labels}
\alias{add_grid_to_map_without_labels}
\title{Add Grid to Map Without Labels}
\usage{
add_grid_to_map_without_labels(my_map)
}
\arguments{
\item{my_map}{An existing leaflet map.}
}
\value{
The updated leaflet map with added grid without labels.
}
\description{
This function adds a grid (graticule) to an existing leaflet map 
without displaying labels.
}
\details{
This function uses the `clearGroup` and `addSimpleGraticule` functions 
from the `leaflet` package to add a grid to the map. The grid is defined by the 
specified interval and includes latitude and longitude lines. Labels for the 
grid lines are not displayed. The grid is added to the "Grid" group on the map.
}
\examples{
\dontrun{
# Example usage
my_map <- init_map_of_sensors(sensors_data)
my_map <- add_grid_to_map_without_labels(my_map)
}

}
