% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{add_north_arrow_to_map}
\alias{add_north_arrow_to_map}
\title{Add North Arrow to Map}
\usage{
add_north_arrow_to_map(my_map)
}
\arguments{
\item{my_map}{An existing leaflet map.}
}
\value{
The updated leaflet map with added north arrow.
}
\description{
This function adds a north arrow marker to an existing leaflet map.
}
\details{
This function uses the `addMarkers` function from the `leaflet` package 
to add a marker representing the north direction to the map. The marker is placed 
at the maximum latitude and longitude coordinates of the map. The icon used for 
the north arrow is specified in the `north_arrow_pin` variable.
}
\examples{
\dontrun{
# Example usage
my_map <- init_map_of_sensors(sensors_data)
my_map <- add_north_arrow_to_map(my_map)
}
}
