% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{add_scale_bar_to_map}
\alias{add_scale_bar_to_map}
\title{Add Scale Bar to Map}
\usage{
add_scale_bar_to_map(my_map)
}
\arguments{
\item{my_map}{An existing leaflet map.}
}
\value{
The updated leaflet map with added scale bar.
}
\description{
This function adds a scale bar to an existing leaflet map.
}
\details{
This function uses the `addScaleBar` function from the `leaflet` package 
to add a scale bar to the bottom right corner of an existing leaflet map. The 
scale bar provides information about the distance on the map in both metric and 
imperial units.
}
\examples{
\dontrun{
# Example usage
my_map <- init_map_of_sensors(sensors_data)
my_map <- add_scale_bar_to_map(my_map)
}
}
