% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{add_sensors_to_map}
\alias{add_sensors_to_map}
\title{Add Sensors to Map}
\usage{
add_sensors_to_map(my_map, sensors)
}
\arguments{
\item{my_map}{An existing leaflet map.}

\item{sensors}{The sensor DataFrame with geospatial information.}
}
\value{
The updated leaflet map with added sensor markers.
}
\description{
This function adds sensors to an existing leaflet map.
}
\details{
This function uses the `addAwesomeMarkers` function from the `leaflet` 
package to add markers for each sensor to an existing leaflet map. The markers 
include information such as the sensor location, icon color, label, and popup 
content with detailed information about the sensor.
}
\examples{
\dontrun{
# Example usage
sensors_data <- read.csv("sensors_geospatial_data.csv")
my_map <- init_map_of_sensors(sensors_data)
my_map <- add_sensors_to_map(my_map, sensors_data)
}
}
\seealso{
\code{\link{sensors_icons_pins}}
}
