% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{add_wind_turbines_to_map}
\alias{add_wind_turbines_to_map}
\title{Add Wind Turbines to Map}
\usage{
add_wind_turbines_to_map(my_map, wind_turbines)
}
\arguments{
\item{my_map}{An existing leaflet map.}

\item{wind_turbines}{The wind turbines DataFrame with geospatial information.}
}
\value{
The updated leaflet map with added wind turbine markers.
}
\description{
This function adds wind turbines to an existing leaflet map.
}
\details{
This function uses the `addMarkers` function from the `leaflet` package 
to add markers for each wind turbine to an existing leaflet map. The markers 
include information such as the wind turbine location, label, and a default icon.
}
\examples{
\dontrun{
# Example usage
turbines_data <- read.csv("wind_turbines_geospatial_data.csv")
my_map <- init_map_of_sensors(sensors_data)
my_map <- add_wind_turbines_to_map(my_map, turbines_data)
}
}
\seealso{
\code{\link{wind_turbine_pin}}
}
