% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{convert_column}
\alias{convert_column}
\title{Convert a Data Column}
\usage{
convert_column(column_data)
}
\arguments{
\item{column_data}{The data of the column to convert.}
}
\value{
The converted values of the column in the appropriate data type.
}
\description{
This function attempts to convert the values of a data column into different 
types (logical, numeric, integer, floating-point, character).
}
\details{
The function identifies non-empty values in the column and then tries to convert 
them into logical, numeric, integer, floating-point, or character types based on certain 
characteristics of the data.
}
\examples{
\dontrun{
# Example usage
data <- c("1", "2", "3", "")
result <- convert_column(data)
}
}
