% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{convert_columns}
\alias{convert_columns}
\title{Convert Columns of a Data Object}
\usage{
convert_columns(df, timestamp_format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{df}{The data object to process.}

\item{timestamp_format}{The format to use for converting the "timestamp" column.}
}
\value{
The data object with converted columns.
}
\description{
This function converts the "timestamp" column to POSIXct type with the 
specified format and attempts to convert other columns into different types.
}
\details{
The function applies the conversion of the "timestamp" column to POSIXct 
type with the specified format. Then, it attempts to convert values in other columns 
into logical, numeric, integer, floating-point, or character types using the 
`convert_column` function.
}
\examples{
\dontrun{
# Example usage
data <- data.frame(timestamp = c("2020-07-19 14:00:00"), value = c("1", "2", "3"))
result <- convert_columns(data)
}
}
