% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{count_common_prefix_elements}
\alias{count_common_prefix_elements}
\title{Count Common Prefix Elements}
\usage{
count_common_prefix_elements(meteo_sensors, strings)
}
\arguments{
\item{meteo_sensors}{A vector of sensor names serving as the reference set.
The function will extract unique prefixes from these names for comparison.}

\item{strings}{A vector of sensor names to be compared with the reference set.}
}
\value{
A data frame where each column corresponds to a unique prefix from the
  reference set, and each row indicates the count of elements in the list
  that share the corresponding prefix.
}
\description{
This function takes a reference set of sensor names and a list of sensor names,
  and counts the number of elements in the list that have a common prefix with each
  element in the reference set.
}
\details{
The function works by converting the reference set of sensor names to a set of unique
  prefixes, and then counting the number of elements in the list that share each prefix.
}
\examples{
\dontrun{
  meteo_sensors <- c("a1", "a2", "a3", "a4", "a5", "lidar", "s1", "s2", "s3")
  input_strings <- c("a3", "a4", "lidar", "s1", "s2")
  result <- count_common_prefix_elements(meteo_sensors, input_strings)
  print(result)
}

}
