% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{ElasticSearch default parameters}
\format{
An object of class \code{list} of length 7.
}
\usage{
default_params
}
\description{
ElasticSearch default parameters
}
\keyword{datasets}
