% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{download_data}
\alias{download_data}
\title{Download Data}
\usage{
download_data(conn, input)
}
\arguments{
\item{conn}{The connection to the remote server.}

\item{input}{A list containing user input parameters, including start and end dates, time sampling, 
and sensor configurations.}
}
\value{
A dataset containing the downloaded data.
}
\description{
Downloads data from a remote server based on user input.
}
\details{
This function retrieves data from a remote server based on the user's input parameters,
  such as start and end dates, time sampling, and sensor configurations. The downloaded data is 
  returned as a dataset.
}
\examples{
conn <- create_connection("http://example.com", port = 9200)
user_input <- list(start_date = "2023-01-01", start_time = "08:00:00",
                   end_date = "2023-01-31", end_time = "18:00:00",
                   time_sampling = "1h", scientific_domain = "physics",
                   acoustic_sensors = c("sensor1", "sensor2"),
                   acoustic_indicators = c("indicator1", "indicator2"),
                   acoustic_frequencies = c(1000, 2000))
dataset <- download_data(conn, user_input)
}
\seealso{
\code{\link{create_connection}}, \code{\link{get_data}}
}
