% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{download_data_body}
\alias{download_data_body}
\title{Download Data Query Body}
\usage{
download_data_body(start_datetime, end_datetime, fields, size = 10000)
}
\arguments{
\item{start_datetime}{The start datetime for the data request.}

\item{end_datetime}{The end datetime for the data request.}

\item{fields}{The fields expected from the database.}

\item{size}{The maximum number of records to retrieve (default is 10000).}
}
\value{
A JSON-formatted query body for use in querying the remote server.
}
\description{
Generates the query body for downloading data from a remote server with timestamp 
filtering.
}
\details{
This function constructs a JSON-formatted query body based on the provided parameters.
  It includes specifications for the time range, aggregation, fields to be retrieved, and other
  relevant settings for querying data from a remote server.
}
\examples{
start_date <- "2023-01-01T00:00:00"
end_date <- "2023-01-31T23:59:59"
fields_to_retrieve <- c("field1", "field2", "field3")
download_data_body(start_date, end_date, fields_to_retrieve)
}
\seealso{
\code{\link{elastic::search}}, \code{\link{sprintf_named}}
}
