% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{es_params}
\alias{es_params}
\title{Try to Load Configuration Parameters}
\format{
An object of class \code{list} of length 7.
}
\usage{
es_params
}
\arguments{
\item{config_file}{The path to the configuration file.}

\item{default_params}{The default parameters to be used if loading fails.}
}
\value{
A list of configuration parameters.
}
\description{
Try to Load Configuration Parameters
}
\examples{
# Example usage:
es_params <- try_load_config(file = "path/to/config.yaml", default_params = list(param1 = 1, param2 = "default"))
}
\keyword{datasets}
