% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{extract_sensor}
\alias{extract_sensor}
\title{Extract Sensor Information from Elasticsearch Index Name}
\usage{
extract_sensor(index_name, field, time_sampling)
}
\arguments{
\item{index_name}{A character string representing the Elasticsearch index name.}

\item{field}{A character string representing the field.}

\item{time_sampling}{A character string representing the time sampling.}
}
\value{
The sensor name.
}
\description{
Extracts sensor information from the given Elasticsearch index name.
}
\examples{
# Example Usage:
extract_sensor("acoustic-a1-10min", "acoustic", "10min")
# Returns: "a1"
}
