% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{format_acoustic_fields_for_elastic}
\alias{format_acoustic_fields_for_elastic}
\title{Format Acoustic Fields for ElasticSearch Query}
\usage{
format_acoustic_fields_for_elastic(acoustic_indicators, acoustic_frequencies)
}
\arguments{
\item{acoustic_indicators}{A vector of acoustic indicators.}

\item{acoustic_frequencies}{A vector of acoustic frequencies.}
}
\value{
A formatted string of fields for the ElasticSearch query.
}
\description{
Formats acoustic indicators and frequencies for an ElasticSearch query.
}
\details{
This function takes vectors of acoustic indicators and frequencies and generates a 
formatted string of fields for an ElasticSearch query.
}
\examples{
indicators <- c("indicator1", "indicator2")
frequencies <- c("freq1", "freq2")
fields_str <- format_acoustic_fields_for_elastic(acoustic_indicators = indicators, acoustic_frequencies = frequencies)
# Result: A formatted string suitable for ElasticSearch query.
}
\seealso{
\code{\link{download_data_body}}, \code{\link{get_elastic_health}}
}
