% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{format_datetime_from_time_slider}
\alias{format_datetime_from_time_slider}
\title{Format Datetime from Time Slider}
\usage{
format_datetime_from_time_slider(the_date, the_time)
}
\arguments{
\item{the_date}{The date selected from the time slider.}

\item{the_time}{The time selected from the time slider.}
}
\value{
A POSIXct object representing the formatted datetime.
}
\description{
Formats a datetime using a date and time from a time slider.
}
\details{
This function takes a date and time from a time slider, combines them into a datetime, 
and returns the result.
}
\examples{
selected_date <- "2020-07-19"
selected_time <- "14:30"
formatted_datetime <- format_datetime_from_time_slider(selected_date, selected_time)
# Result: POSIXct object representing the formatted datetime.
}
\seealso{
\code{\link{download_data}}, \code{\link{get_data}}
}
