% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{format_datetimes_for_csv}
\alias{format_datetimes_for_csv}
\title{Format Datetimes for CSV}
\usage{
format_datetimes_for_csv(
  start_date_times,
  out_format = "\%Y-\%m-\%d_\%H:\%M:\%OS"
)
}
\arguments{
\item{start_date_times}{A vector of date-time strings to be formatted.}

\item{out_format}{The desired output format for the date-time strings. Defaults to \%Y-\%m-\%d_\%H:\%M:\%OS.}
}
\value{
A list of formatted date-time strings.
}
\description{
Formats a vector of date-time strings for CSV output.
}
\details{
This function takes a vector of date-time strings, interprets them as UTC, 
and formats them for CSV output. The user can specify the desired output format using 
the `out_format` parameter.
}
\examples{
date_times <- c("2023-01-01T08:00:00Z", "2023-01-02T12:30:00Z")
formatted_dates <- format_datetimes_for_csv(date_times)
# Result: List of formatted date-time strings for CSV.
}
\seealso{
\code{\link{download_data}}, \code{\link{get_data}}
}
