% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{format_datetimes_for_elastic}
\alias{format_datetimes_for_elastic}
\title{Format Datetimes for ElasticSearch}
\usage{
format_datetimes_for_elastic(start_date, start_time, end_date, end_time)
}
\arguments{
\item{start_date}{The start date provided by the user.}

\item{start_time}{The start time provided by the user.}

\item{end_date}{The end date provided by the user.}

\item{end_time}{The end time provided by the user.}
}
\value{
A list containing formatted start and end datetimes.
}
\description{
Formats user-provided dates and times into the required format for 
querying ElasticSearch.
}
\details{
This function takes user-provided start and end dates along with 
corresponding start and end times and formats them into the required format 
for querying ElasticSearch. The formatted datetimes are returned as a list.
}
\examples{
start_date <- "2023-01-01"
start_time <- "08:00:00"
end_date <- "2023-01-31"
end_time <- "18:00:00"
formatted_datetimes <- format_datetimes_for_elastic(start_date, start_time, end_date, end_time)
# Result: List with formatted start and end datetimes.
}
\seealso{
\code{\link{download_data}}, \code{\link{get_data}}
}
