% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{format_elapsed_time}
\alias{format_elapsed_time}
\title{Format Elapsed Time}
\usage{
format_elapsed_time(elapsed_time)
}
\arguments{
\item{elapsed_time}{The elapsed time in seconds.}
}
\value{
A formatted string indicating the elapsed time in minutes and seconds.
}
\description{
Formats elapsed time into a human-readable string.
}
\details{
This function takes an elapsed time in seconds and formats it into a human-readable string. 
If the elapsed time is greater than or equal to one minute, it is displayed in minutes and seconds; 
otherwise, it is displayed in seconds.
}
\examples{
elapsed_seconds <- 75
formatted_time <- format_elapsed_time(elapsed_time = elapsed_seconds)
# Result: "1 min 15.00 s"
}
