% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_common_fields_variables}
\alias{get_common_fields_variables}
\title{Get Common Fields/Variables}
\usage{
get_common_fields_variables(mappings_list)
}
\arguments{
\item{mappings_list}{A list of mappings.}
}
\value{
A vector containing the common fields/variables.
}
\description{
This function takes a list of mappings and finds the common fields/variables.
}
\details{
Get Common Fields/Variables
This function takes a list of mappings and finds the common fields/variables
}
\examples{
mappings_list <- list(
  c("field_A", "field_B", "field_C"),
  c("field_B", "field_C", "field_D"),
  c("field_A", "field_B", "field_D")
)
common_fields <- get_common_fields_variables(mappings_list)
print(common_fields)
}
