% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_data}
\alias{get_data}
\title{Retrieve Data from Elasticsearch}
\usage{
get_data(
  conn,
  start_date,
  start_time,
  end_date,
  end_time,
  time_sampling,
  scientific_domain,
  acoustic_sensors,
  acoustic_indicators,
  acoustic_frequencies,
  meteo_sensors,
  meteo_variables,
  size = 10000
)
}
\arguments{
\item{conn}{The Elasticsearch connection.}

\item{start_date}{The start date for data retrieval.}

\item{start_time}{The start time for data retrieval.}

\item{end_date}{The end date for data retrieval.}

\item{end_time}{The end time for data retrieval.}

\item{time_sampling}{The time sampling interval.}

\item{scientific_domain}{The scientific domain for data retrieval.}

\item{acoustic_sensors}{The acoustic sensors for data retrieval.}

\item{acoustic_indicators}{The acoustic indicators for data retrieval.}

\item{acoustic_frequencies}{The acoustic frequencies for data retrieval.}

\item{size}{The number of data samples to retrieve in each request.}
}
\value{
A dataframe containing the retrieved data.
}
\description{
This function retrieves data from Elasticsearch based on specified parameters.
}
\details{
This function constructs Elasticsearch queries based on the specified parameters 
and retrieves data in chunks to handle pagination.
}
\examples{
\dontrun{
# Example usage
conn <- elasticsearch::connect(host = "localhost", port = 9200)
start_date <- "2023-01-01"
start_time <- "12:00:00"
end_date <- "2023-01-02"
end_time <- "12:00:00"
time_sampling <- "1h"
scientific_domain <- "domain1"
acoustic_sensors <- "sensor1"
acoustic_indicators <- c("indicator1", "indicator2")
acoustic_frequencies <- c("frequency1", "frequency2")
size <- 10000
data <- get_data(conn, start_date, start_time, end_date, end_time, time_sampling, 
scientific_domain, acoustic_sensors, acoustic_indicators, acoustic_frequencies, size)
}
}
