% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_elastic_health}
\alias{get_elastic_health}
\title{Get ElasticSearch Health}
\usage{
get_elastic_health(conn)
}
\arguments{
\item{conn}{The ElasticSearch connection object.}
}
\value{
A data frame containing various health indicators from ElasticSearch.
}
\description{
Retrieves the health status information from an ElasticSearch connection.
}
\details{
This function sends a request to ElasticSearch to retrieve health status
  information and parses the response into a structured data frame.
}
\examples{
conn <- elastic::connect()
get_elastic_health(conn)
}
\seealso{
\code{\link{elastic::connect}}, \code{\link{elastic::cat_health}}
}
