% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_indexes_info}
\alias{get_indexes_info}
\title{Get Elasticsearch Index Information}
\usage{
get_indexes_info(conn)
}
\arguments{
\item{conn}{An Elasticsearch connection object.}
}
\value{
A data frame containing information about each index.
}
\description{
This function extracts index information from the output of elastic::cat_indices().
}
\examples{
\dontrun{
conn <- connect("http://your-elasticsearch-host:9200")
index_info <- get_indexes_info(conn)
}
}
