% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_indexes_mappings}
\alias{get_indexes_mappings}
\title{Get Indexes Mappings}
\usage{
get_indexes_mappings(conn, indexes)
}
\arguments{
\item{conn}{An Elasticsearch connection object.}

\item{indexes}{A character vector containing the names of the Elasticsearch indexes.}
}
\value{
A list containing the flattened mappings for each index.
}
\description{
This function retrieves the mappings of selected Elasticsearch indexes.
}
\details{
Get mappings of selected indexes
}
\examples{
\dontrun{
indexes_mappings <- get_indexes_mappings(conn = elastic_conn, indexes = c("index1", "index2"))
}
}
