% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_indexes_names}
\alias{get_indexes_names}
\title{Get Indexes Names for a Specific Field}
\usage{
get_indexes_names(indexes_info, field)
}
\arguments{
\item{indexes_info}{A data frame containing information about indexes.}

\item{field}{The field for which index names are to be retrieved.}
}
\value{
A character vector containing the names of indexes for the specified field.
}
\description{
This function retrieves the names of indexes for a specified field, excluding the "extraneous_noise" sensor.
}
\examples{
\dontrun{
# Assuming indexes_info is a data frame with appropriate structure
indexes_names <- get_indexes_names(indexes_info, "acoustic")
}
}
\seealso{
\code{\link{elastic::cat_indices}} for information about the indexes_info data frame structure.
}
