% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_meteo_variables}
\alias{get_meteo_variables}
\title{Get Meteo Variables
This function extracts meteo sensors and related variables from a list of meteo mappings.}
\usage{
get_meteo_variables(mappings_list)
}
\arguments{
\item{mappings_list}{A list of meteo mappings.}
}
\value{
A list containing meteo sensors and their related variables.
}
\description{
Get Meteo Variables
This function extracts meteo sensors and related variables from a list of meteo mappings.
}
\examples{
mappings_list <- list("meteo-a4-10min" = c("Temp", "Hum"), "meteo-s3-1min" = c("Temp", "Wind_Speed"))
meteo_variables <- get_meteo_variables(mappings_list)
}
