% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_sensors_for_field}
\alias{get_sensors_for_field}
\title{Get a list of sensors for a specific field, optionally avoiding certain sensors.}
\usage{
get_sensors_for_field(indexes_info, wanted_field, avoided_sensors = NULL)
}
\arguments{
\item{indexes_info}{A data frame containing Elasticsearch index information.}

\item{wanted_field}{The desired field for which sensors are to be retrieved.}

\item{avoided_sensors}{A character vector of sensors to avoid (default is NULL).}
}
\value{
A character vector containing the list of sensors for the specified field.
}
\description{
This function extracts a list of sensors for a given field from Elasticsearch index information.
Optionally, you can provide a list of sensors to avoid.
}
\examples{
get_sensors_for_field(indexes_info, wanted_field = 'acoustic', avoided_sensors = 'extraneous_noise')
get_sensors_for_field(indexes_info, wanted_field = 'meteo', avoided_sensors = c('extraneous_noise', 'a1', 'a2', 'a3'))
}
