% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_sensors_names_starting_with}
\alias{get_sensors_names_starting_with}
\title{Get Sensors Names Starting With
This function filters a list of sensor names to include only those starting with a specified pattern.}
\usage{
get_sensors_names_starting_with(sensors, names_starting_with)
}
\arguments{
\item{sensors}{A character vector containing sensor names.}

\item{names_starting_with}{A character string specifying the pattern that sensor names should start with.}
}
\value{
A character vector of sensor names starting with the specified pattern.
}
\description{
Get Sensors Names Starting With
This function filters a list of sensor names to include only those starting with a specified pattern.
}
\examples{
sensors <- c("a1", "a2", "a3", "b1", "b2", "b3")
selected_sensors <- get_sensors_names_starting_with(sensors, "a")
}
