% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_timesampling_unit_and_value}
\alias{get_timesampling_unit_and_value}
\title{Get Timesampling Unit and Value}
\usage{
get_timesampling_unit_and_value(time_sampling)
}
\arguments{
\item{time_sampling}{A string representing the timesampling, e.g., "1h", "30s", "2d", etc.}
}
\value{
A named list containing numeric value and unit.
}
\description{
Extracts numeric value and unit from a timesampling string.
}
\details{
This function parses a timesampling string (e.g., "1h", "30s") and extracts the 
numeric value and unit.
}
\examples{
timesampling_str <- "1h"
timesampling_info <- get_timesampling_unit_and_value(timesampling_str)
# Result: List with "num" and "unit" components.
}
\seealso{
\code{\link{format_datetime_from_time_slider}}, \code{\link{download_data}}
}
