% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_wind_turbine_relative_coords}
\alias{get_wind_turbine_relative_coords}
\title{Get Wind Turbine Relative Coordinates}
\usage{
get_wind_turbine_relative_coords(ref_point_coords)
}
\arguments{
\item{ref_point_coords}{A dataframe with latitude and longitude of the reference point.}
}
\value{
A Simple Features (sf) object representing wind turbines with relative coordinates.
}
\description{
This function creates a dataset of wind turbines with relative 
coordinates calculated with respect to a reference point.
}
\details{
This function creates a dataset of wind turbines with fixed coordinates. It 
then calculates the relative coordinates of each wind turbine with respect to a reference 
point specified by `ref_point_coords` and returns the result as an sf object.
}
\examples{
\dontrun{
# Example usage
ref_point_coords <- data.frame(latitude = ..., longitude = ...)  # Replace with your reference point coordinates
wind_turbines <- get_wind_turbine_relative_coords(ref_point_coords)
}
}
