% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{Global Functions}
\alias{Global Functions}
\title{global.R}
\description{
Functions and Methods used in the Shinydashboard scripts
}
\details{
Current date: `r format(Sys.Date(), "%B %e, %Y")`
}
\section{libraries}{
 Libraries
}

