% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{init_map_of_sensors}
\alias{init_map_of_sensors}
\title{Initialize Map of Sensors}
\usage{
init_map_of_sensors(sensors, wind_turbines)
}
\arguments{
\item{sensors}{The sensor geospatial DataFrame.}
}
\value{
An initialized leaflet map.
}
\description{
This function initializes a leaflet map centered around the geographical coordinates of the sensors.
}
\details{
This function uses the `leaflet` package to create a leaflet map centered around the geographical coordinates 
of the sensors. The center and zoom level are determined based on the calculated centre point of the sensor locations.
}
\examples{
\dontrun{
# Example usage
sensors_data <- read.csv("sensors_geospatial_data.csv")
my_map <- init_map_of_sensors(sensors_data)
}
}
\seealso{
\code{\link{get_geodf_centre_point}}
}
