% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{Initializations}
\alias{Initializations}
\alias{title}
\title{Shinydashboard script for PIBE Database Management Dashboard
-----------------------------------------------------------
This script defines a Shiny dashboard with various tabs for managing and visualizing data related to the PIBE project.
It utilizes Shinydashboard, leaflet, and plotly for interactive elements.
-----------------------------------------------------------}
\format{
An object of class \code{character} of length 1.
}
\usage{
title
}
\arguments{
\item{title}{The title of the dashboard.}

\item{logo}{The path to the logo image.}
}
\description{
Load the required libraries for the dashboard.

## Initializations

This part initialize the title and logo for the dashboard.
}
\section{Initializations and loading Libraries}{
NA
}

\keyword{datasets}
