% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{modal.options}
\alias{modal.options}
\title{Shiny Modal Options Function}
\usage{
modal.options(input, output, session, options)
}
\arguments{
\item{input}{The input object from the Shiny app.}

\item{output}{The output object from the Shiny app.}

\item{session}{The Shiny session object.}

\item{options}{A list containing the modal options.}
}
\value{
A modified list of options based on user interactions.
}
\description{
This function creates a Shiny modal with options and handles user interactions.
}
\examples{
# Example usage:
options <- modal.options(input, output, session, list(check1 = FALSE))
}
