% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{modal.options.UI}
\alias{modal.options.UI}
\title{Shiny Modal Options UI Function}
\usage{
modal.options.UI(id)
}
\arguments{
\item{id}{The ID of the modal options.}
}
\value{
A Shiny UI element containing an action button to show the modal.
}
\description{
This function creates a Shiny modal options UI element.
}
\examples{
# Example usage:
ui <- modal.options.UI("my_modal")
}
