% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{my_options}
\alias{my_options}
\title{Global options change to show milliseconds}
\format{
An object of class \code{list} of length 3.
}
\usage{
my_options
}
\description{
Global options change to show milliseconds
}
\keyword{datasets}
