% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{order_indicators}
\alias{order_indicators}
\title{Order Acoustic Indicators}
\usage{
order_indicators(x)
}
\arguments{
\item{x}{A character vector of strings to be sorted.}
}
\value{
A numeric vector indicating the order of elements in the sorted array.
}
\description{
This function is designed for sorting strings based on the presence of digits.
Strings containing digits are considered greater and placed at the end, while
strings without digits are considered smaller and placed at the beginning.
}
\examples{
strings <- c("L10", "L50", "L90", "Leq")
sorted_strings <- strings[order(sapply(strings, indicators_custom_order), strings)]
print(sorted_strings)
}
\seealso{
order
}
