% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{ping_database}
\alias{ping_database}
\title{Ping Database}
\usage{
ping_database(conn)
}
\arguments{
\item{conn}{The connection object to the remote server.}
}
\value{
A list containing the ping result and, if successful, a message with connection details.
}
\description{
Sends a ping request to check the status of the connection to a remote server.
}
\details{
This function sends a ping request to the specified remote server and measures
  the time taken for the server to respond. If the ping is successful, it returns a list
  with the ping result and a message indicating the successful connection.
}
\examples{
conn <- elastic::connect()
ping_database(conn)
}
\seealso{
\code{\link{elastic::connect}}, \code{\link{milliseconds}}
}
