% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{sensor_xlsx_file_to_df}
\alias{sensor_xlsx_file_to_df}
\title{Convert Excel File to Data Frame}
\usage{
sensor_xlsx_file_to_df(xlsx_filename, sheetname)
}
\arguments{
\item{file_path}{Path to the Excel file (.xlsx) containing sensor data.}

\item{sheet_name}{Name of the worksheet in the Excel file to extract (optional,
by default, the first sheet will be used).}
}
\value{
A data frame representing the sensor data extracted from the Excel file.
}
\description{
This function converts an Excel file (.xlsx) containing sensor data into a data frame.
It takes the path to the Excel file and an optional sheet name as input and returns
a data frame representing the sensor data.
}
\section{functions_sensors}{
 Sensors utilities
}

\examples{
# Example usage with the file "sensor_data.xlsx" and the sheet "SensorData"
df <- sensor_xlsx_file_to_df(file_path = "sensor_data.xlsx", sheet_name = "SensorData")

# Display the first few rows of the data frame
head(df)
}
