% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{sensors_icons_pins}
\alias{sensors_icons_pins}
\title{Sensors Icons Pins}
\usage{
sensors_icons_pins()
}
\value{
A list of Awesome Icons for acoustic, mast, and LIDAR sensors with black, green, and red colors.
}
\description{
This function generates a list of Awesome Icons for different sensor types and colors.
}
\details{
This function uses the `awesomeIconList` function from the `leaflet` package to create icons 
for different sensor types (acoustic, mast, LIDAR) with variations in color (black, green, red).
}
\examples{
\dontrun{
# Example usage
icons_list <- sensors_icons_pins()
}
}
