% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny server function for PIBE project presentation}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{The input values from the Shiny app interface.}

\item{output}{Reactive expressions and outputs to update in the Shiny app.}

\item{session}{The Shiny session object.}
}
\description{
This function defines the server logic for the PIBE project presentation Shiny app.
}
\details{
server.R

This server function includes sections for PIBE project presentation, session timer,
connection to the server, sensor map, data download, data visualization, and miscellaneous.
}
\examples{
server(input, output, session)

PIBE PROJECT PRESENTATION

PIBE project logo

PIBE project html document

Session timer

Session counter

REMOTE SERVER CONNECTION

Close shiny app and window

Connection details to the Elasticsearch engine

Ping connection


Connection Button

Cluster health

Server health

SENSOR MAP

Locations of sensors
Reference sensor point
Set coordinates relative to mast
Wind turbines location

Sensor map marker click event

Map rendering on PIBE project presentation tab

Map rendering on data download tab


Reset map view (centre of dataframe)

Table of sensors

DATA DOWNLOAD

Update the values of the dates and times input
Acoustic variables
Meteo variables

Calculating download time

Full path of output file

Acoustic data plot






#'
if (input$jitter)
  p <- p + geom_jitter()
#' 
if (input$smooth)
  p <- p + geom_smooth()
#' 
height <- session$clientData$output_p_height
width <- session$clientData$output_p_width
ggplotly(p, height = height, width = width, )

Table of data

DATA VISUALIZATION

facetColumns <- reactive({
  choices <- c(None='.', str_remove_all(names(mydata$plot), c("timestamp|value")))
  return(choices)
})
#' 
output$facet_columns <- renderUI({
  selectInput('facet_col', 'Facet Column', facetColumns())
})


Time slider update

Date range update


Data refresh

Sensor map rendering

Table of acoustic data

MISCELLANEOUS


Reuse the same output binding on different tab panels
}
