% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{sessions_secs_timer}
\alias{sessions_secs_timer}
\title{Time Elapsed Since Session Started}
\usage{
sessions_secs_timer(start_datetime)
}
\arguments{
\item{start_datetime}{The timestamp representing the start of the session.}
}
\value{
The time elapsed in seconds since the session started.
}
\description{
Calculates the time elapsed in seconds since the specified session start time.
}
\details{
This function takes the start timestamp of a session and calculates the
  time elapsed in seconds up to the current moment.
}
\section{functions_server}{
 Remote server connection utilities

This documentation provides details about the remote server connection utilities.
}

\examples{
start_time <- as.POSIXct("2023-01-01 12:00:00", tz = "UTC")
sessions_secs_timer(start_time)
}
\seealso{
\code{\link{as.POSIXct}}, \code{\link{difftime}}, \code{\link{round}}, \code{\link{str_remove}}
}
