% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_sensors_color_icon}
\alias{set_sensors_color_icon}
\title{Set Sensors Color and Icon}
\usage{
set_sensors_color_icon(sensors_df, selected_sensors)
}
\arguments{
\item{sensors_df}{A dataframe containing sensor information.}

\item{selected_sensors}{A character vector specifying the selected sensor names.}
}
\value{
A modified dataframe with updated color and icon information for selected 
and non-selected sensors.
}
\description{
This function sets the color and icon for selected and non-selected sensors.
}
\details{
This function sets the color of selected sensors to green and non-selected sensors 
to red. It also updates the "color_icon" column based on the sensor locations and selected colors.
}
\examples{
\dontrun{
# Example usage
sensors_data <- ...  # Replace with your sensor dataframe
selected_sensor_names <- c("sensor1", "sensor2")  # Replace with your selected sensor names
updated_sensors <- set_sensors_color_icon(sensors_data, selected_sensors = selected_sensor_names)
}
}
