% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{Shinydashboard UI}
\alias{Shinydashboard UI}
\title{Header Section}
\arguments{
\item{title}{The title of the dashboard.}

\item{logo}{The path to the logo image.


### Title

### Message menu

### Notifications

### Tasks

### Quit app}

\item{id}{The ID of the sidebar.}
}
\description{
This section defines the Shinydashboard UI, including the header, sidebar, and body sections.

This tab presents the PIBE project and the WP2 package.

This section defines the sidebar of the dashboard.

This section defines the body of the dashboard.

This tab presents the PIBE project and the WP2 package.
Show html files presenting the PIBE project and the WP2 package
Table of sensors

### Data download

This tab allows users to download data based on specified criteria.
#### Acoustic sensor(s)
#### Acoustic indicator(s)
#### Frequency-ies
#### Meteo sensor(s)

This tab provides visualizations based on the selected data.


#### Time series

This sub-section provides a time series plot.



#### Scatter plot

This sub-section provides a scatter plot.
}
\section{Shinydashboard UI}{
NA
}

\seealso{
[Shinydashboard documentation](https://rstudio.github.io/shinydashboard/)

[Leaflet documentation](https://rstudio.github.io/leaflet/)

[Plotly documentation](https://plotly.com/r/shinyapp-dashboards/)

[Shinydashboard documentation](https://rstudio.github.io/shinydashboard/)


### PIBE project

### Data download

### Data visualization

[Shinydashboard documentation](https://rstudio.github.io/shinydashboard/)


### PIBE project presentation
}
