% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{sprintf_named}
\alias{sprintf_named}
\title{Formatted sprintf with Named Arguments}
\usage{
sprintf_named(fmt, ...)
}
\arguments{
\item{fmt}{A character string containing the format specification.}

\item{...}{Named arguments to be substituted into the format string.}
}
\value{
A character vector resulting from applying the formatted string.
}
\description{
This function extends the base R `sprintf` function by allowing
  the use of named arguments in the format string.
}
\details{
This function provides a convenient way to use named arguments in a
  sprintf-like format string. The named arguments are specified using the
  `%{name}` syntax, where `name` corresponds to the name of the argument.
}
\examples{
fmt <- "Hello, \%{name}! You are \%{age} years old."
sprintf_named(fmt, name = "John", age = 30)
}
\seealso{
\code{\link{sprintf}}
}
