% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{unique_field_sensor_variable_triplets}
\alias{unique_field_sensor_variable_triplets}
\title{Unique Field, Sensor, and Variable Triplets}
\usage{
unique_field_sensor_variable_triplets(df)
}
\arguments{
\item{df}{A dataframe containing columns "field", "sensor", and "variable".}
}
\value{
A tibble with unique triplets of "field", "sensor", and "variable". The three
values are concatenated into a string with "." as separators.
}
\description{
This function takes a dataframe as input and returns a tibble with unique triplets
of "field", "sensor", and "variable". The three values are concatenated into a
string with "." as separators.
}
\details{
The function uses the distinct function to obtain unique combinations of
"field", "sensor", and "variable". It then uses mutate with paste to create a
new column "field.sensor.var" with the concatenated string.
}
\examples{
\dontrun{
# Create a sample dataframe
df <- tibble::tibble(
  field = c("acoustic", "acoustic", "meteo"),
  sensor = c("a1", "a2", "lidar"),
  variable = c("Leq_A", "Leq_B", "Hum")
)
# Call the function
unique_triplets <- unique_field_sensor_variable_triplets(df)
}
}
