% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{wind_turbine_pin}
\alias{wind_turbine_pin}
\title{Define map icons}
\format{
An object of class \code{leaflet_icon} of length 3.
}
\usage{
wind_turbine_pin
}
\description{
Define map icons
}
\keyword{datasets}
